+++ src/Amadeus/Client/RequestCreator/Converter/Ticket/ATCShopperMasterPricerTravelBoardSearchConv.php
@@ -20,20 +20,27 @@
  * @license https://opensource.org/licenses/Apache-2.0 Apache 2.0
  */
 
-namespace Amadeus\Client\RequestOptions;
+namespace Amadeus\Client\RequestCreator\Converter\Ticket;
+
+use Amadeus\Client\RequestCreator\Converter\BaseConverter;
+use Amadeus\Client\RequestOptions\TicketAtcShopperMpTbSearchOptions;
+use Amadeus\Client\Struct;
 
 /**
- * TicketCreateTsmFareElOptions
+ * ATCShopperMasterPricerTravelBoardSearchConv
  *
- * @package Amadeus\Client\RequestOptions
+ * @package Amadeus\Client\RequestCreator\Converter\Ticket
  * @author Dieter Devlieghere <dieter.devlieghere@benelux.amadeus.com>
  */
-class TicketCreateTsmFareElOptions extends TicketDisplayTsmFareElOptions
+class ATCShopperMasterPricerTravelBoardSearchConv extends BaseConverter
 {
     /**
-     * Unstructured fare element information
-     *
-     * @var string
+     * @param TicketAtcShopperMpTbSearchOptions $requestOptions
+     * @param int|string $version
+     * @return Struct\Ticket\AtcShopperMasterPricerTravelBoardSearch
      */
-    public $info;
+    public function convert($requestOptions, $version)
+    {
+        return new Struct\Ticket\AtcShopperMasterPricerTravelBoardSearch($requestOptions);
+    }
 }